refactor: separated out ability upgrades into its own file and added types
