--- src/parser/parsers/abilities.py
@@ -1,208 +1,0 @@
-import parser.maps as maps
-import utils.json_utils as json_utils
-import utils.num_utils as num_utils
-from loguru import logger
-
-
-class AbilityParser:
-    def __init__(self, abilities_data, heroes_data, localizations):
-        self.abilities_data = abilities_data
-        self.heroes_data = heroes_data
-        self.localizations = localizations
-
-    def run(self):
-        all_abilities = {}
-        ability_key = ''
-        try:
-            for ability_key in self.abilities_data:
-                ability = self._parse_ability(ability_key)
-                if ability:
-                    all_abilities[ability_key] = ability
-
-            return all_abilities
-        except Exception as e:
-            logger.error(f'Failed to parse ability {ability_key} - {e}')
-            raise e
-
-    def _parse_ability(self, ability_key):
-        ability = self.abilities_data[ability_key]
-
-        if type(ability) is not dict:
-            return
-
-        if 'm_eAbilityType' not in ability:
-            return
-
-        if ability['m_eAbilityType'] not in ['EAbilityType_Innate', 'EAbilityType_Signature', 'EAbilityType_Ultimate']:
-            return
-
-        ability_data = {
-            'Key': ability_key,
-            'Name': self.localizations.get(ability_key, None),
-            'IsDisabled': ability.get('m_bDisabled', False),
-        }
-
-        stats = ability.get('m_mapAbilityProperties', {})
-        for key in stats:
-            stat = stats[key]
-            value = self._get_stat_value(key, stat)
-            scale = self._get_scale(stat)
-            if scale:
-                ability_data[key] = {'Value': value, 'Scale': scale}
-            else:
-                ability_data[key] = value
-
-        if 'm_vecAbilityUpgrades' in ability:
-            ability_data['Upgrades'] = self._parse_upgrades(ability)
-        else:
-            ability_data['Upgrades'] = []
-
-        formatted_ability_data = {}
-        for attr_key, attr_value in ability_data.items():
-            formatted_ability_data[attr_key] = num_utils.remove_uom(attr_value)
-
-        return json_utils.sort_dict(formatted_ability_data)
-
-    def _parse_upgrades(self, ability):
-        upgrade_sets = ability['m_vecAbilityUpgrades']
-        parsed_upgrade_sets = []
-
-        for upgrade_set in upgrade_sets:
-            upgrades = upgrade_set.get('m_vecPropertyUpgrades')
-            if upgrades is None:
-                continue
-
-            # 1. Extract raw data from each upgrade entry
-            raw_upgrades = []
-            for upgrade in upgrades:
-                prop = upgrade.get('m_strPropertyName')
-                raw_value = upgrade.get('m_strBonus')
-                upgrade_type = upgrade.get('m_eUpgradeType')
-                scale_type = upgrade.get('m_eScaleStatFilter')
-
-                value = num_utils.assert_number(raw_value)
-
-                stat = json_utils.deep_get(ability, 'm_mapAbilityProperties', prop)
-                if stat:
-                    value = self._convert_stat(stat, raw_value, value)
-
-                raw_upgrades.append(
-                    {
-                        'prop': prop,
-                        'value': value,
-                        'upgrade_type': upgrade_type,
-                        'scale_type': scale_type,
-                    }
-                )
-
-            # 2. Group by prop so we can reason about base + scales together
-            upgrades_by_prop = {}
-            for raw in raw_upgrades:
-                upgrades_by_prop.setdefault(raw['prop'], []).append(raw)
-
-            # 3. Build the parsed upgrade set from grouped data
-            parsed_upgrade_set = {}
-            for prop, entries in upgrades_by_prop.items():
-                base_value = 0
-                scales = []
-
-                for entry in entries:
-                    if entry['upgrade_type'] in ['EAddToBase', None]:
-                        base_value = entry['value']
-                    elif entry['upgrade_type'] in ['EAddToScale', 'EMultiplyScale']:
-                        scale_type = entry['scale_type']
-                        if scale_type is None:
-                            base_stat = json_utils.deep_get(ability, 'm_mapAbilityProperties', prop)
-                            if base_stat and 'm_subclassScaleFunction' in base_stat:
-                                scale_func = base_stat['m_subclassScaleFunction']
-                                scale_type = scale_func.get('m_eSpecificStatScaleType')
-                                if scale_type is None and 'tech' in scale_func.get('_class', ''):
-                                    scale_type = 'ETechPower'
-                        scales.append(
-                            {
-                                'Value': entry['value'],
-                                'Type': maps.get_scale_type(scale_type),
-                            }
-                        )
-
-                if scales:
-                    # if there are multiple scales for a prop, output as an array
-                    if len(scales) > 1:
-                        parsed_upgrade_set[prop] = {'Value': base_value, 'Scale': scales}
-                    else:
-                        parsed_upgrade_set[prop] = {'Value': base_value, 'Scale': scales[0]}
-                else:
-                    parsed_upgrade_set[prop] = base_value
-
-            parsed_upgrade_sets.append(parsed_upgrade_set)
-
-        return parsed_upgrade_sets
-
-    def _get_stat_value(self, key, stat):
-        value = None
-
-        if 'm_strValue' in stat:
-            value = stat['m_strValue']
-        elif 'm_strVAlue' in stat:
-            value = stat['m_strVAlue']
-        else:
-            return
-
-        return self._convert_stat(stat, key, value)
-
-    def _convert_stat(self, stat: dict, key: str, value: str | int | float):
-        """Convert a value to the correct unit of measure and format based on the raw stat's information
-
-        Args:
-            stat (dict): Raw ability stat which includes information on its units and class
-            key (str): Key of ability stat
-            value (str | int | float): Value of ability stat
-
-        Returns:
-            int | float: Converted value
-        """
-        # if the value ends with "m", it is already converted to the correct units
-        if isinstance(value, str) and value.endswith('m'):
-            return num_utils.assert_number(value[:-1])
-
-        # specific to ChannelMoveSpeed, a "-1" indicates stationary, so no need to convert units
-        if key == 'ChannelMoveSpeed' and value == '-1':
-            return -1
-
-        units = stat.get('m_eDisplayUnits')
-        strClass = stat.get('m_strCSSClass')
-
-        # some ranges are written as "1500 2000" to denote a specific range
-        if isinstance(value, str) and strClass == 'range':
-            ranges = value.split(' ')
-            if len(ranges) == 2:
-                lower = num_utils.assert_number(ranges[0])
-                upper = num_utils.assert_number(ranges[1])
-                if units in ['EDisplayUnit_Meters', 'EDisplayUnit_MetersPerSecond']:
-                    return f'{num_utils.convert_engine_units_to_meters(lower)} {num_utils.convert_engine_units_to_meters(upper)}'
-                else:
-                    return f'{lower} {upper}'
-
-        value = num_utils.assert_number(value)
-
-        if units in ['EDisplayUnit_Meters', 'EDisplayUnit_MetersPerSecond']:
-            return num_utils.convert_engine_units_to_meters(value)
-
-        return value
-
-    def _get_scale(self, stat):
-        """
-        Get scale data for the ability attribute, which will refer to how the value of the attribute
-        scales with another stat, usually Spirit
-        """
-        if 'm_subclassScaleFunction' in stat:
-            scale = stat['m_subclassScaleFunction']
-            # Only include scale with a value, as not sure what
-            # any others mean so far.
-            if 'm_flStatScale' in scale:
-                return {
-                    'Value': scale['m_flStatScale'],
-                    'Type': maps.get_scale_type(scale.get('m_eSpecificStatScaleType', 'ETechPower')),
-                }
-
-        return
